S15.c: decompose inference-router/src/routes/inference.rs (1359 → 776 LOC)
Extract the 582-line `chat_completions` handler (POST /v1/chat/completions)
into a new sibling module `routes/chat_completions.rs`. Registered as a
private mod in `routes/mod.rs`; `inference_routes()` imports it via
`use super::chat_completions::chat_completions;`.

`inference.rs` retains the route-builder fns (inference_routes,
foundry_agent_routes, foundry_standalone_routes) and the smaller
handlers (completions, responses, embeddings, images_generations,
images_generations_v1, list_models, list_deployments, foundry_proxy).

Verification:
- inference.rs: 1359 → 776 LOC (under §15 800-LOC cap)
- cargo build / clippy --all-targets -- -D warnings / fmt --check: clean
- cargo test --lib: 608 passed; 0 failed

No behavioral change: handler body moved verbatim; visibility raised
from `async fn` to `pub(super) async fn` so the parent module's
route-builder can register it.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-inference-router-routes.md
