+++ inference-router/src/routes/mod.rs
@@ -44,6 +44,8 @@ pub use egress::egress_routes;
 mod inference;
 pub use inference::{foundry_agent_routes, foundry_standalone_routes, inference_routes};
 
+mod chat_completions;
+
 mod mcp;
 pub use mcp::{
     MCP_SESSION_HEADER, McpRouteState, mcp_route, platform_mcp_route, protected_mcp_route,
